Front-load Debian packaging deps into docker build
In an effort to speed up debian compilation and make things more
predictable, I tried to accomplish three things here:
* Remove apt logic from ansible and put into docker

Ansible is incredibly slow at handling this AND if we want to
parallelize we cant have each container doing the exact same apt work.
Thats just silly.

* Peg apt dependencies to specific versions

This makes packages that we produce more reliable and easier to QA.
There is an added burden that we have to check for upstream updates
however.

* Parallelize the debian build process so its more efficient

I added a container per package and passed a setting to ansible to force
it to run each host as fast as possible.